% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{cal_p}
\alias{cal_p}
\title{Calculate the significance of the dynamic signal for a single gene's single time expression.
Compared to baseline NB confidence interval.}
\usage{
cal_p(obs.val, size, mu)
}
\arguments{
\item{obs.val, }{estimated value from non-baseline model.}

\item{size, }{the size of negative binomial model.}

\item{mu, }{the estimated mean of constant negative binomial model}
}
\value{
a numeric p-value for time t.
}
\description{
Calculate the significance of the dynamic signal for a single gene's single time expression.
Compared to baseline NB confidence interval.
}
