% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{cal_time_p_single_gene}
\alias{cal_time_p_single_gene}
\title{Calculate the significance of the dynamic signal for a single gene's over alll the time points.}
\usage{
cal_time_p_single_gene(const.output, spline.output)
}
\arguments{
\item{const.output, }{list returned from fit_single_gene_const function.}

\item{spline.output, }{dataframe returned from fit_single_gene_spline function.}
}
\value{
p-value for a single point.
}
\description{
Calculate the significance of the dynamic signal for a single gene's over alll the time points.
}
