% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{combine_p_single_gene}
\alias{combine_p_single_gene}
\title{Combine multiple p-value using Fisher's p-value combination method.}
\usage{
combine_p_single_gene(p.arr)
}
\arguments{
\item{p.arr, }{a vector of p-values.}
}
\value{
a numeric combined p-value.
}
\description{
Combine multiple p-value using Fisher's p-value combination method.
}
