% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{curveFitting.new}
\alias{curveFitting.new}
\title{CurveFitting function}
\usage{
curveFitting.new(perm.dat.1, points)
}
\arguments{
\item{perm.dat.1, }{data to fit the curve.}

\item{points, }{array of time values.}
}
\value{
fitted count over time
}
\description{
To perform Local Polynomial Regression Fitting
}
