% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CurveComp_Perm.R
\name{draw_CurveComp_Perm}
\alias{draw_CurveComp_Perm}
\title{Draw DDEGs trajectories from two master.list object and also show LOESS curve fitting with permuataon}
\usage{
draw_CurveComp_Perm(
  master.list.1,
  master.list.2,
  ht.1,
  pathway,
  group.1.name,
  group.2.name,
  n.perm = 500,
  parall = F,
  pvalue.threshold = 0.05
)
}
\description{
Draw DDEGs trajectories from two master.list object and also show LOESS curve fitting with permuataon
}
