% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_GeneTraj.R
\name{draw_GeneTraj}
\alias{draw_GeneTraj}
\title{Draw gene(s) trajectory with observed data and fitted data}
\usage{
draw_GeneTraj(
  master.list,
  gene.symbol.arr,
  savepdf.path = NA,
  ncol = 5,
  nrow = 3,
  fig.width = 15,
  fig.height = 10
)
}
\arguments{
\item{master.list, }{a list object. Output from the run_TrendCatcher with ID conversion to add Symbol column to master table.}

\item{gene.symbol.arr, }{a character array. It must be a subset of row names from the master.list$master.table$Symbol.
The Symbol column need get_GeneEnsembl2Symbol function to convert original ensembl ID into gene symbol.}

\item{savepdf.path, }{an obsolute file path to save the figure as PDF file. By default is NA, it will be printed.}

\item{ncol, }{an integer variable. If more than one gene need to be plotted, it will layout as grid structure.
This represents the number of column of the grid layout.}

\item{nrow, }{an integer variable. If more than one gene need to be plotted, it will layout as grid structure.
This represents the number of row of the grid layout.}

\item{fig.width, }{a numeric variable. If save figure as PDF file, the width of the PDF file. By default is 15.}

\item{fig.height, }{a numeric variable. If save figure as PDF file, the height of the PDF file. By default is 10.}
}
\value{
"arrangelist" "list" object.
}
\description{
This function takes the master.list object output from run_TrendCatcher function, and an array of gene(s).
It will draw gene(s) trajectory with observed data and fitted data.
}
