% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_TimeHeatmap_selGO.R
\name{draw_TimeHeatmap_selGO}
\alias{draw_TimeHeatmap_selGO}
\title{Subset TimeHeatmap}
\usage{
draw_TimeHeatmap_selGO(
  time_heatmap,
  sel.go,
  master.list,
  GO.perc.thres = 0,
  nDDEG.thres = 0,
  term.width = 80,
  figure.title = "",
  save.tiff.path = NA,
  tiff.res = 100,
  tiff.width = 1500,
  tiff.height = 1500
)
}
\description{
Subset TimeHeatmap
}
