% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_TrajClusterGrid.R
\name{draw_TrajClusterGrid}
\alias{draw_TrajClusterGrid}
\title{Draw Grouped DDEGs Trajectories in Grid Plot}
\usage{
draw_TrajClusterGrid(
  master.list,
  min.traj.n = 10,
  save.as.PDF = NA,
  pdf.width = 10,
  pdf.height = 10
)
}
\arguments{
\item{master.list, }{a list object. The output from run_TrendCatcher function, contains master.table element.}

\item{min.traj.n, }{an integer variable. The minimum number of genes from the same sub-type trajectory. By default if 10.}

\item{save.as.PDF, }{a string. The absolute file path to save the figure as PDF file if needed. If set to NA, will print
the figure instead of saving it as PDF file.}

\item{pdf.width, }{a numeric variable. The PDF file width size. By default is 10.}

\item{pdf.height, }{a numeric variable. The PDF file height size. By default is 10.}
}
\description{
Group all DDEGs based on their sub-type trajectory patterns and plot their trajectories together,
then layout all sub-type trajectory patterns which contains more than N genes in a grid plot.
Each individual sub-grid plot is titled with sub-type trajectory pattern and number of genes included.
X-axis is the time, Y-aixs is log2 transformed fitted count trajectory.
}
