% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_TrajClusterPie.R
\name{draw_TrajClusterPie}
\alias{draw_TrajClusterPie}
\title{Draw Grouped DDEGs Main-type and Sub-type Composition in Hierarchical Pie Chart}
\usage{
draw_TrajClusterPie(
  master.list,
  fig.title = "",
  inner.radius = 0.7,
  cex.out = 1,
  cex.in = 1
)
}
\arguments{
\item{master.list, }{a list object. The output from run_TrendCatcher function, contains master.table element.}

\item{fig.title, }{a string. The main title of the figure. By default if "".}

\item{inner.radius, }{a numeric variable. The inner pie chart radius size. By default is 0.7.}

\item{cex.out, }{a numeric variable. The text size of label of outer pie chart. By default is 1.}

\item{cex.in, }{a numeric variable. The text size of the label of inner pie chart. By default is 1.}
}
\description{
Group all DDEGs based on their main-type and sub-type trajectory patterns and plot their composition in a
hierachical pie chart. Inner pie chart represents the main-type trajectory pattern composition. The outer pie chart
represents sub-type trajectory pattern composition.
}
