% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{fit_single_gene_const}
\alias{fit_single_gene_const}
\title{Fit the baseline count data into a constant negative binomial model.}
\usage{
fit_single_gene_const(
  count.arr,
  disp.var,
  MAXIT = 1000,
  RELTOL = 10^(-8),
  trace = 10
)
}
\arguments{
\item{count.arr, }{a data frame returned from transform_single_gene_df, only with the baseline time.}

\item{disp.var, }{the dispersion value estimated from DESeq2.}

\item{MAXIT, }{\enumerate{
\item 
}}

\item{RELTOL, }{10*-8}

\item{trace, }{10}
}
\value{
a list contain all the estimated value from NB model.
}
\description{
Fit the baseline count data into a constant negative binomial model.
}
