% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get_rep_array}
\alias{get_rep_array}
\title{Get the replicate array from count table}
\usage{
get_rep_array(raw.count.df)
}
\arguments{
\item{raw.count.df, }{a count table.}
}
\value{
a numeric array of each sample's replicate.
}
\description{
It take the count table, grep the replicate element from column name. The column name of the count table must satisfy "Prj_Time_Rep1" format.
}
