% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GeneSymbol.R
\name{get_Symbol2GeneEnsembl}
\alias{get_Symbol2GeneEnsembl}
\title{ID convention (from SYMBOL to ENSEMBL)}
\usage{
get_Symbol2GeneEnsembl(symbol.arr, dataset = "mmusculus_gene_ensembl")
}
\arguments{
\item{symbol.arr, }{a character array. The SYMBOL array.}

\item{dataset, }{must be either "mmusculus_gene_ensembl" or "hsapiens_gene_ensembl".}
}
\value{
A data frame contains 3 columns, "Gene", original ID, "Symbol" and "description".
}
\description{
This function takes an array of SYMBOL and convert it into GENE ENSEMBL.
}
