% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{transform_single_gene_df}
\alias{transform_single_gene_df}
\title{Convert a single gene's count row number into data frame with two columns.}
\usage{
transform_single_gene_df(gene.row.info, gene.name, time.arr, rep.arr)
}
\arguments{
\item{gene.row.info, }{a single row from count table.}

\item{gene.name, }{the gene name.}

\item{time.arr, }{the return value from get_time_array function.}

\item{rep.arr, }{the return value from get_rep_array function.}
}
\value{
a dataframe object ordered by time and replicate id.
}
\description{
Convert a single gene's count row number into data frame with two columns.
}
