% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CurveComp.R
\name{draw_CurveComp}
\alias{draw_CurveComp}
\title{Draw DDEGs trajectories from two master.list object and also show LOESS curve fitting}
\usage{
draw_CurveComp(
  master.list.1,
  master.list.2,
  ht.1,
  pathway,
  group.1.name,
  group.2.name
)
}
\arguments{
\item{master.list.1, }{a list object. The output from run_TrendCatcher function, contains master.table element.}

\item{master.list.2, }{a list object. The output from run_TrendCatcher function, contains master.table element.}

\item{ht.1, }{TimeHeatmap object. The output from draw_TimeHeatmap_GO function, contains  GO.df object.}

\item{pathway, }{characters. Must be a biological pathway from GO.df, Description column.}

\item{group.1.name, }{characters. For example, severe group.}

\item{group.2.name, }{characters. For example, moderate group.}
}
\value{
A ggplot object and plot.
}
\description{
This funcitons takes the master.list output from run_TrendCatcher, and merge.df output from draw_TimeHeatmap_GO and draw_TimeHeatmap_enrichR.
And showing all the genes used for enrichment analysis and their logFC compared to previous break point.
}
